requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-dotenv==0.10.1
pytz==2018.9
six==1.12.0
Werkzeug==0.15.1
